more refactoring of sync tests
