+++ K6/api/tests/register/er-sync/testcase_2_add_styremedlem.js
@@ -3,34 +3,30 @@ import { EnterpriseTokenGenerator } from "../../../../common-imports.js";
 import { AuthorizedPartiesClient, RegisterApiClient } from "../../../../clients/authentication/index.js";
 import { GetAuthorizedParties } from "../../../building-blocks/authentication/authorized-parties/index.js";
 import { SubmitErData } from "../../../building-blocks/register/index.js";
-import { generateOrgNr } from "../../../../helpers.js";
+import { generateOrgNr, retry } from "../../../../helpers.js";
 import { runErSyncTestcase } from "./helper.js";
 
 /**
- * @file remove-medl.js
- * @description Verifies that removing a MEDL (Styremedlem) in ER is correctly
+ * @file testcase_2_add_styremedlem.js
+ * @description Verifies that adding a MEDL (Styremedlem) in ER is correctly
  * synced to Altinn Register and reflected in authorized parties.
  *
- * k6 run remove-medl.js \
+ * k6 run testcase_2_add_styremedlem.js \
  *   -e ENVIRONMENT=at22 -e BASE_URL=https://platform.at22.altinn.cloud \
  *   -e SOAP_ER_USERNAME=<u> -e SOAP_ER_PASSWORD=<p> \
  *   -e REGISTER_SUBSCRIPTION_KEY=<key>
  *
- * @requires ENV.ENVIRONMENT                - Target environment (e.g. tt02, at22)
- * @requires ENV.BASE_URL                   - Base URL for the Register API
- * @requires ENV.SOAP_ER_USERNAME           - Username for the ER SOAP API
- * @requires ENV.SOAP_ER_PASSWORD           - Password for the ER SOAP API
- * @requires ENV.REGISTER_SUBSCRIPTION_KEY  - Subscription key for the Register API
+ * @see README.md
  */
 
 export const options = {
     scenarios: {
-        "remove-board-member": { executor: "shared-iterations", exec: "removeMedl", vus: 1, iterations: 1 },
+        "add-board-member": { executor: "shared-iterations", exec: "addMedl", vus: 1, iterations: 1 },
     },
 };
 
-const DAGL = { fnr: "26827896992", fornavn: "VIKTIG", slektsnavn: "ORIDÉ" };
-const MEDL = { fnr: "10921148513", fornavn: "UKLAR",  slektsnavn: "PLAST" };
+const DAGL     = { fnr: "09861798434", fornavn: "AKADEMISK", slektsnavn: "HAKE" };
+const NEW_MEDL = { fnr: "10921148513", fornavn: "UKLAR",     slektsnavn: "PLAST" };
 
 function buildPrepXml(orgNr) {
     return `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -41,11 +37,11 @@ function buildPrepXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00250" mottaker="ALT" type="A" />
+            <head avsender="ER" dato="20260512" kjoerenr="00240" mottaker="ALT" type="A" />
             <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="N" undersakstype="NY" foersteOverfoering="J" datoFoedt="20200101" datoSistEndret="20260512">
                 <infotype felttype="NAVN" endringstype="N">
-                    <navn1>REMOVE MEDL TEST AS</navn1>
-                    <rednavn>REMOVE MEDL TEST AS</rednavn>
+                    <navn1>ADD MEDL TEST AS</navn1>
+                    <rednavn>ADD MEDL TEST AS</rednavn>
                 </infotype>
                 <infotype felttype="FADR" endringstype="N">
                     <postnr>0150</postnr>
@@ -59,17 +55,6 @@ function buildPrepXml(orgNr) {
                     <kommunenr>0301</kommunenr>
                     <adresse1>Testveien 10</adresse1>
                 </infotype>
-                <samendringer data="D" felttype="MEDL" endringstype="N" type="R">
-                    <rolleFratraadt>N</rolleFratraadt>
-                    <rolleRekkefoelge>1</rolleRekkefoelge>
-                    <rolleFoedselsnr>${MEDL.fnr}</rolleFoedselsnr>
-                    <fornavn>${MEDL.fornavn}</fornavn>
-                    <slektsnavn>${MEDL.slektsnavn}</slektsnavn>
-                    <postnr>0150</postnr>
-                    <adresse1>Testveien 12</adresse1>
-                    <adresseLandkode>NO</adresseLandkode>
-                    <personstatus>L</personstatus>
-                </samendringer>
                 <samendringer data="D" felttype="DAGL" endringstype="N" type="R">
                     <rolleFratraadt>N</rolleFratraadt>
                     <rolleRekkefoelge>1</rolleRekkefoelge>
@@ -93,7 +78,7 @@ export function setup() {
     return { orgNr: generateOrgNr() };
 }
 
-export function removeMedl({ orgNr = generateOrgNr() } = {}) {
+export function addMedl({ orgNr = generateOrgNr() } = {}) {
     const prep = buildPrepXml(orgNr);
 
     const change = `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -104,10 +89,18 @@ export function removeMedl({ orgNr = generateOrgNr() } = {}) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00340" mottaker="ALT" type="A" />
-            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
-                <samendringer data="D" felttype="MEDL" endringstype="U" type="R">
-                    <rolleFoedselsnr>${MEDL.fnr}</rolleFoedselsnr>
+            <head avsender="ER" dato="20260512" kjoerenr="00330" mottaker="ALT" type="A" />
+            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="NY" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
+                <samendringer data="D" felttype="MEDL" endringstype="N" type="R">
+                    <rolleFratraadt>N</rolleFratraadt>
+                    <rolleRekkefoelge>1</rolleRekkefoelge>
+                    <rolleFoedselsnr>${NEW_MEDL.fnr}</rolleFoedselsnr>
+                    <fornavn>${NEW_MEDL.fornavn}</fornavn>
+                    <slektsnavn>${NEW_MEDL.slektsnavn}</slektsnavn>
+                    <postnr>0150</postnr>
+                    <adresse1>Testveien 12</adresse1>
+                    <adresseLandkode>NO</adresseLandkode>
+                    <personstatus>L</personstatus>
                 </samendringer>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
@@ -123,18 +116,27 @@ export function removeMedl({ orgNr = generateOrgNr() } = {}) {
     const apClient = new AuthorizedPartiesClient(__ENV.BASE_URL, new EnterpriseTokenGenerator(tokenOpts));
 
     runErSyncTestcase(
-        "Remove board member (MEDL)",
+        "Add board member (MEDL)",
         prep,
         change,
         orgNr,
-        { "org is accessible in Register after MEDL removed": (p) => p.partyType === "organization" },
+        { "org is accessible in Register after MEDL added": (p) => p.partyType === "organization" },
         {
             afterChange: () => {
-                group("Verify - MEDL no longer has access to org", () => {
-                    const parties = GetAuthorizedParties(apClient, "urn:altinn:person:identifier-no", MEDL.fnr, { includeAltinn2: false, includePartiesViaKeyRoles: true });
-                    check(parties, {
-                        [`MEDL (${MEDL.fornavn} ${MEDL.slektsnavn}) no longer has access to org`]: (p) =>
-                            Array.isArray(p) && !p.some((party) => party.organizationNumber === orgNr || party.orgNumber === orgNr),
+                group("Verify - new MEDL has access to org", () => {
+                    let verifiedParties = null;
+                    retry(
+                        () => {
+                            const parties = GetAuthorizedParties(apClient, "urn:altinn:person:identifier-no", NEW_MEDL.fnr, { includeAltinn2: false, includePartiesViaKeyRoles: true });
+                            if (!Array.isArray(parties)) return false;
+                            const hasAccess = parties.some((p) => p.organizationNumber === orgNr || p.orgNumber === orgNr);
+                            if (hasAccess) verifiedParties = parties;
+                            return hasAccess;
+                        },
+                        { retries: 15, intervalSeconds: 20, testscenario: "add-board-member - new MEDL access" },
+                    );
+                    check(verifiedParties, {
+                        [`new MEDL (${NEW_MEDL.fornavn} ${NEW_MEDL.slektsnavn}) has access to org`]: (p) => p !== null,
                     });
                 });
             },
@@ -159,10 +161,13 @@ function buildCleanupXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00411" mottaker="ALT" type="A" />
+            <head avsender="ER" dato="20260512" kjoerenr="00410" mottaker="ALT" type="A" />
             <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
                 <samendringer data="D" felttype="DAGL" endringstype="U" type="R">
                     <rolleFoedselsnr>${DAGL.fnr}</rolleFoedselsnr>
+                </samendringer>
+                <samendringer data="D" felttype="MEDL" endringstype="U" type="R">
+                    <rolleFoedselsnr>${NEW_MEDL.fnr}</rolleFoedselsnr>
                 </samendringer>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
