+++ K6/api/tests/register/er-sync/testcase_5_replace_styreleder.js
@@ -1,36 +1,35 @@
-import { group } from "k6";
-import { generateOrgNr } from "../../../../helpers.js";
-import { runErSyncTestcase } from "./helper.js";
-import { RegisterApiClient } from "../../../../clients/authentication/index.js";
+import { group, check } from "k6";
+import { EnterpriseTokenGenerator } from "../../../../common-imports.js";
+import { AuthorizedPartiesClient, RegisterApiClient } from "../../../../clients/authentication/index.js";
+import { GetAuthorizedParties } from "../../../building-blocks/authentication/authorized-parties/index.js";
 import { SubmitErData } from "../../../building-blocks/register/index.js";
+import { generateOrgNr, retry } from "../../../../helpers.js";
+import { runErSyncTestcase } from "./helper.js";
 
 /**
- * @file change-fadr.js
- * @description Verifies that a change to FADR (Forretningsadresse) in ER is correctly
- * synced to Altinn Register.
+ * @file change-styr.js
+ * @description Verifies that a change to LEDE (Styreleder) in ER is correctly
+ * synced to Altinn Register and reflected in authorized parties.
  *
- * k6 run change-fadr.js \
+ * k6 run change-styr.js \
  *   -e ENVIRONMENT=at22 -e BASE_URL=https://platform.at22.altinn.cloud \
  *   -e SOAP_ER_USERNAME=<u> -e SOAP_ER_PASSWORD=<p> \
  *   -e REGISTER_SUBSCRIPTION_KEY=<key>
  *
- * @requires ENV.ENVIRONMENT                - Target environment (e.g. tt02, at22)
- * @requires ENV.BASE_URL                   - Base URL for the Register API
- * @requires ENV.SOAP_ER_USERNAME           - Username for the ER SOAP API
- * @requires ENV.SOAP_ER_PASSWORD           - Password for the ER SOAP API
- * @requires ENV.REGISTER_SUBSCRIPTION_KEY  - Subscription key for the Register API
+ * @see README.md
  */
 
 export const options = {
     scenarios: {
-        "change-business-address": { executor: "shared-iterations", exec: "fadrChange", vus: 1, iterations: 1 },
+        "replace-styreleder": { executor: "shared-iterations", exec: "styrChange", vus: 1, iterations: 1 },
     },
 };
 
-const LEDE  = { fnr: "02831899053", fornavn: "LILLA",     slektsnavn: "NETTADRESSE" };
-const MEDL  = { fnr: "03823648714", fornavn: "SEIN",      slektsnavn: "ELV" };
-const MEDL2 = { fnr: "26812099719", fornavn: "STOR",      slektsnavn: "KAPPE" };
-const DAGL  = { fnr: "05830299450", fornavn: "SPENNENDE", slektsnavn: "BRØKSTREK" };
+const OLD_STYRELEDER = { fnr: "16918598441", fornavn: "TREG", slektsnavn: "HUNKATT" };
+const NEW_STYRELEDER = { fnr: "20845996750", fornavn: "TYPISK", slektsnavn: "TIMEBU" };
+const STYREMEDLEM = { fnr: "57925901581", fornavn: "PASSIV", slektsnavn: "EKORNHALE" };
+const DAGLIG_LEDER = { fnr: "18914598245", fornavn: "UTÅLMODIG", slektsnavn: "MASKIN" };
+
 
 function buildPrepXml(orgNr) {
     return `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -41,11 +40,11 @@ function buildPrepXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00210" mottaker="ALT" type="A" />
+            <head avsender="ER" dato="20260512" kjoerenr="00230" mottaker="ALT" type="A" />
             <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="N" undersakstype="NY" foersteOverfoering="J" datoFoedt="20200101" datoSistEndret="20260512">
                 <infotype felttype="NAVN" endringstype="N">
-                    <navn1>FADR CHANGE TEST AS</navn1>
-                    <rednavn>FADR CHANGE TEST AS</rednavn>
+                    <navn1>BYTTE STYRELEDER AS</navn1>
+                    <rednavn>BYTTE STYRELEDER AS</rednavn>
                 </infotype>
                 <infotype felttype="FADR" endringstype="N">
                     <postnr>0150</postnr>
@@ -62,9 +61,9 @@ function buildPrepXml(orgNr) {
                 <samendringer data="D" felttype="LEDE" endringstype="N" type="R">
                     <rolleFratraadt>N</rolleFratraadt>
                     <rolleRekkefoelge>1</rolleRekkefoelge>
-                    <rolleFoedselsnr>${LEDE.fnr}</rolleFoedselsnr>
-                    <fornavn>${LEDE.fornavn}</fornavn>
-                    <slektsnavn>${LEDE.slektsnavn}</slektsnavn>
+                    <rolleFoedselsnr>${OLD_STYRELEDER.fnr}</rolleFoedselsnr>
+                    <fornavn>${OLD_STYRELEDER.fornavn}</fornavn>
+                    <slektsnavn>${OLD_STYRELEDER.slektsnavn}</slektsnavn>
                     <postnr>0150</postnr>
                     <adresse1>Testveien 11</adresse1>
                     <adresseLandkode>NO</adresseLandkode>
@@ -73,31 +72,20 @@ function buildPrepXml(orgNr) {
                 <samendringer data="D" felttype="MEDL" endringstype="N" type="R">
                     <rolleFratraadt>N</rolleFratraadt>
                     <rolleRekkefoelge>1</rolleRekkefoelge>
-                    <rolleFoedselsnr>${MEDL.fnr}</rolleFoedselsnr>
-                    <fornavn>${MEDL.fornavn}</fornavn>
-                    <slektsnavn>${MEDL.slektsnavn}</slektsnavn>
+                    <rolleFoedselsnr>${STYREMEDLEM.fnr}</rolleFoedselsnr>
+                    <fornavn>${STYREMEDLEM.fornavn}</fornavn>
+                    <slektsnavn>${STYREMEDLEM.slektsnavn}</slektsnavn>
                     <postnr>0150</postnr>
                     <adresse1>Testveien 12</adresse1>
-                    <adresseLandkode>NO</adresseLandkode>
-                    <personstatus>L</personstatus>
-                </samendringer>
-                <samendringer data="D" felttype="MEDL" endringstype="N" type="R">
-                    <rolleFratraadt>N</rolleFratraadt>
-                    <rolleRekkefoelge>2</rolleRekkefoelge>
-                    <rolleFoedselsnr>${MEDL2.fnr}</rolleFoedselsnr>
-                    <fornavn>${MEDL2.fornavn}</fornavn>
-                    <slektsnavn>${MEDL2.slektsnavn}</slektsnavn>
-                    <postnr>0150</postnr>
-                    <adresse1>Testveien 13</adresse1>
                     <adresseLandkode>NO</adresseLandkode>
                     <personstatus>L</personstatus>
                 </samendringer>
                 <samendringer data="D" felttype="DAGL" endringstype="N" type="R">
                     <rolleFratraadt>N</rolleFratraadt>
                     <rolleRekkefoelge>1</rolleRekkefoelge>
-                    <rolleFoedselsnr>${DAGL.fnr}</rolleFoedselsnr>
-                    <fornavn>${DAGL.fornavn}</fornavn>
-                    <slektsnavn>${DAGL.slektsnavn}</slektsnavn>
+                    <rolleFoedselsnr>${DAGLIG_LEDER.fnr}</rolleFoedselsnr>
+                    <fornavn>${DAGLIG_LEDER.fornavn}</fornavn>
+                    <slektsnavn>${DAGLIG_LEDER.slektsnavn}</slektsnavn>
                     <postnr>0150</postnr>
                     <adresse1>Testveien 14</adresse1>
                     <adresseLandkode>NO</adresseLandkode>
@@ -115,7 +103,7 @@ export function setup() {
     return { orgNr: generateOrgNr() };
 }
 
-export function fadrChange({ orgNr = generateOrgNr() } = {}) {
+export function styrChange({ orgNr = generateOrgNr() } = {}) {
     const prep = buildPrepXml(orgNr);
 
     const change = `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -126,14 +114,17 @@ export function fadrChange({ orgNr = generateOrgNr() } = {}) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00306" mottaker="ALT" type="A" />
-            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
-                <infotype felttype="FADR" endringstype="N">
-                    <postnr>0350</postnr>
-                    <landkode>NO</landkode>
-                    <kommunenr>0301</kommunenr>
-                    <adresse1>Ny Forretningsgate 5</adresse1>
-                </infotype>
+            <head avsender="ER" dato="20260512" kjoerenr="00322" mottaker="ALT" type="A" />
+            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E"
+            undersakstype="NY" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
+                <samendringer data="D" felttype="LEDE" endringstype="U" type="R">
+                    <rolleFoedselsnr>${OLD_STYRELEDER.fnr}</rolleFoedselsnr>
+                </samendringer>
+                <samendringer data="D" felttype="LEDE" endringstype="N" type="R">
+                    <rolleFoedselsnr>${NEW_STYRELEDER.fnr}</rolleFoedselsnr>
+                    <fornavn>${NEW_STYRELEDER.fornavn}</fornavn>
+                    <slektsnavn>${NEW_STYRELEDER.slektsnavn}</slektsnavn>
+                </samendringer>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
         </batchAjourholdXML>]]></ns:ERData>
@@ -141,14 +132,45 @@ export function fadrChange({ orgNr = generateOrgNr() } = {}) {
     </soapenv:Body>
 </soapenv:Envelope>`;
 
+    const tokenOpts = new Map();
+    tokenOpts.set("env", __ENV.ENVIRONMENT);
+    tokenOpts.set("ttl", 3600);
+    tokenOpts.set("scopes", "altinn:accessmanagement/authorizedparties.resourceowner");
+    const apClient = new AuthorizedPartiesClient(__ENV.BASE_URL, new EnterpriseTokenGenerator(tokenOpts));
+
     runErSyncTestcase(
-        "Change business address (FADR)",
+        "Replace Styreleder (LEDE)",
         prep,
         change,
         orgNr,
+        { "org is accessible in Register after Styreleder is replaced": (p) => p.partyType === "organization" },
         {
-            "org businessAddress updated to Ny Forretningsgate 5": (p) => p.businessAddress?.address === "Ny Forretningsgate 5",
-            "org businessAddress postalCode updated to 0350": (p) => p.businessAddress?.postalCode === "0350",
+            afterChange: () => {
+                group("Verify - new STYR has access to org", () => {
+                    let verifiedParties = null;
+                    retry(
+                        () => {
+                            const parties = GetAuthorizedParties(apClient, "urn:altinn:person:identifier-no", NEW_STYRELEDER.fnr, { includeAltinn2: false, includePartiesViaKeyRoles: true });
+                            if (!Array.isArray(parties)) return false;
+                            const hasAccess = parties.some((p) => p.organizationNumber === orgNr || p.orgNumber === orgNr);
+                            if (hasAccess) verifiedParties = parties;
+                            return hasAccess;
+                        },
+                        { retries: 15, intervalSeconds: 20, testscenario: "replace-styreleder - new STYR access" },
+                    );
+                    check(verifiedParties, {
+                        [`new Styreleder (${NEW_STYRELEDER.fornavn} ${NEW_STYRELEDER.slektsnavn}) has access to org`]: (p) => p !== null,
+                    });
+                });
+
+                group("Verify - old STYR no longer has access to org", () => {
+                    const parties = GetAuthorizedParties(apClient, "urn:altinn:person:identifier-no", OLD_STYRELEDER.fnr, { includeAltinn2: false, includePartiesViaKeyRoles: true });
+                    check(parties, {
+                        [`old STYR (${OLD_STYRELEDER.fornavn} ${OLD_STYRELEDER.slektsnavn}) no longer has access to org`]: (p) =>
+                            Array.isArray(p) && !p.some((party) => party.organizationNumber === orgNr || party.orgNumber === orgNr),
+                    });
+                });
+            },
         },
     );
 
@@ -170,19 +192,16 @@ function buildCleanupXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00404" mottaker="ALT" type="A" />
+            <head avsender="ER" dato="20260512" kjoerenr="00401" mottaker="ALT" type="A" />
             <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
                 <samendringer data="D" felttype="LEDE" endringstype="U" type="R">
-                    <rolleFoedselsnr>${LEDE.fnr}</rolleFoedselsnr>
+                    <rolleFoedselsnr>${NEW_STYRELEDER.fnr}</rolleFoedselsnr>
                 </samendringer>
                 <samendringer data="D" felttype="MEDL" endringstype="U" type="R">
-                    <rolleFoedselsnr>${MEDL.fnr}</rolleFoedselsnr>
-                </samendringer>
-                <samendringer data="D" felttype="MEDL" endringstype="U" type="R">
-                    <rolleFoedselsnr>${MEDL2.fnr}</rolleFoedselsnr>
+                    <rolleFoedselsnr>${STYREMEDLEM.fnr}</rolleFoedselsnr>
                 </samendringer>
                 <samendringer data="D" felttype="DAGL" endringstype="U" type="R">
-                    <rolleFoedselsnr>${DAGL.fnr}</rolleFoedselsnr>
+                    <rolleFoedselsnr>${DAGLIG_LEDER.fnr}</rolleFoedselsnr>
                 </samendringer>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
