+++ K6/api/tests/register/er-sync/testcase_7_update_contact_info.js
@@ -1,36 +1,30 @@
-import { group, check } from "k6";
-import { EnterpriseTokenGenerator } from "../../../../common-imports.js";
-import { AuthorizedPartiesClient, RegisterApiClient } from "../../../../clients/authentication/index.js";
-import { GetAuthorizedParties } from "../../../building-blocks/authentication/authorized-parties/index.js";
-import { SubmitErData } from "../../../building-blocks/register/index.js";
-import { generateOrgNr, retry } from "../../../../helpers.js";
+import { group } from "k6";
+import { generateOrgNr } from "../../../../helpers.js";
 import { runErSyncTestcase } from "./helper.js";
+import { RegisterApiClient } from "../../../../clients/authentication/index.js";
+import { SubmitErData } from "../../../building-blocks/register/index.js";
 
 /**
- * @file add-medl.js
- * @description Verifies that adding a MEDL (Styremedlem) in ER is correctly
- * synced to Altinn Register and reflected in authorized parties.
+ * @file change-contact.js
+ * @description Verifies that changes to contact information (TFON, TFAX, EPOS, IADR) in ER
+ * are correctly synced to Altinn Register.
  *
- * k6 run add-medl.js \
+ * k6 run change-contact.js \
  *   -e ENVIRONMENT=at22 -e BASE_URL=https://platform.at22.altinn.cloud \
  *   -e SOAP_ER_USERNAME=<u> -e SOAP_ER_PASSWORD=<p> \
  *   -e REGISTER_SUBSCRIPTION_KEY=<key>
  *
- * @requires ENV.ENVIRONMENT                - Target environment (e.g. tt02, at22)
- * @requires ENV.BASE_URL                   - Base URL for the Register API
- * @requires ENV.SOAP_ER_USERNAME           - Username for the ER SOAP API
- * @requires ENV.SOAP_ER_PASSWORD           - Password for the ER SOAP API
- * @requires ENV.REGISTER_SUBSCRIPTION_KEY  - Subscription key for the Register API
+ * @see README.md
  */
 
 export const options = {
     scenarios: {
-        "add-board-member": { executor: "shared-iterations", exec: "addMedl", vus: 1, iterations: 1 },
+        "update-contact-info": { executor: "shared-iterations", exec: "contactChange", vus: 1, iterations: 1 },
     },
 };
 
-const DAGL     = { fnr: "09861798434", fornavn: "AKADEMISK", slektsnavn: "HAKE" };
-const NEW_MEDL = { fnr: "10921148513", fornavn: "UKLAR",     slektsnavn: "PLAST" };
+const LEDE = { fnr: "28824198537", fornavn: "SNÅL",       slektsnavn: "LERKE" };
+const DAGL = { fnr: "57896202792", fornavn: "INTERESSERT", slektsnavn: "FANGE" };
 
 function buildPrepXml(orgNr) {
     return `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -41,11 +35,11 @@ function buildPrepXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00240" mottaker="ALT" type="A" />
+            <head avsender="ER" dato="20260512" kjoerenr="00229" mottaker="ALT" type="A" />
             <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="N" undersakstype="NY" foersteOverfoering="J" datoFoedt="20200101" datoSistEndret="20260512">
                 <infotype felttype="NAVN" endringstype="N">
-                    <navn1>ADD MEDL TEST AS</navn1>
-                    <rednavn>ADD MEDL TEST AS</rednavn>
+                    <navn1>CONTACT CHANGE TEST AS</navn1>
+                    <rednavn>CONTACT CHANGE TEST AS</rednavn>
                 </infotype>
                 <infotype felttype="FADR" endringstype="N">
                     <postnr>0150</postnr>
@@ -59,6 +53,29 @@ function buildPrepXml(orgNr) {
                     <kommunenr>0301</kommunenr>
                     <adresse1>Testveien 10</adresse1>
                 </infotype>
+                <infotype felttype="TFON" endringstype="N">
+                    <opplysning>22334455</opplysning>
+                </infotype>
+                <infotype felttype="TFAX" endringstype="N">
+                    <opplysning>22334456</opplysning>
+                </infotype>
+                <infotype felttype="EPOS" endringstype="N">
+                    <opplysning>contact-test@example.com</opplysning>
+                </infotype>
+                <infotype felttype="IADR" endringstype="N">
+                    <opplysning>http://initial.example.com</opplysning>
+                </infotype>
+                <samendringer data="D" felttype="LEDE" endringstype="N" type="R">
+                    <rolleFratraadt>N</rolleFratraadt>
+                    <rolleRekkefoelge>1</rolleRekkefoelge>
+                    <rolleFoedselsnr>${LEDE.fnr}</rolleFoedselsnr>
+                    <fornavn>${LEDE.fornavn}</fornavn>
+                    <slektsnavn>${LEDE.slektsnavn}</slektsnavn>
+                    <postnr>0150</postnr>
+                    <adresse1>Testveien 11</adresse1>
+                    <adresseLandkode>NO</adresseLandkode>
+                    <personstatus>L</personstatus>
+                </samendringer>
                 <samendringer data="D" felttype="DAGL" endringstype="N" type="R">
                     <rolleFratraadt>N</rolleFratraadt>
                     <rolleRekkefoelge>1</rolleRekkefoelge>
@@ -82,7 +99,7 @@ export function setup() {
     return { orgNr: generateOrgNr() };
 }
 
-export function addMedl({ orgNr = generateOrgNr() } = {}) {
+export function contactChange({ orgNr = generateOrgNr() } = {}) {
     const prep = buildPrepXml(orgNr);
 
     const change = `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -93,19 +110,20 @@ export function addMedl({ orgNr = generateOrgNr() } = {}) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00330" mottaker="ALT" type="A" />
-            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="NY" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
-                <samendringer data="D" felttype="MEDL" endringstype="N" type="R">
-                    <rolleFratraadt>N</rolleFratraadt>
-                    <rolleRekkefoelge>1</rolleRekkefoelge>
-                    <rolleFoedselsnr>${NEW_MEDL.fnr}</rolleFoedselsnr>
-                    <fornavn>${NEW_MEDL.fornavn}</fornavn>
-                    <slektsnavn>${NEW_MEDL.slektsnavn}</slektsnavn>
-                    <postnr>0150</postnr>
-                    <adresse1>Testveien 12</adresse1>
-                    <adresseLandkode>NO</adresseLandkode>
-                    <personstatus>L</personstatus>
-                </samendringer>
+            <head avsender="ER" dato="20260512" kjoerenr="00321" mottaker="ALT" type="A" />
+            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
+                <infotype felttype="TFON" endringstype="N">
+                    <opplysning>99887711</opplysning>
+                </infotype>
+                <infotype felttype="TFAX" endringstype="N">
+                    <opplysning>99887712</opplysning>
+                </infotype>
+                <infotype felttype="EPOS" endringstype="N">
+                    <opplysning>contact-test-oppdatert@example.com</opplysning>
+                </infotype>
+                <infotype felttype="IADR" endringstype="N">
+                    <opplysning>http://oppdatert.example.com</opplysning>
+                </infotype>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
         </batchAjourholdXML>]]></ns:ERData>
@@ -113,37 +131,16 @@ export function addMedl({ orgNr = generateOrgNr() } = {}) {
     </soapenv:Body>
 </soapenv:Envelope>`;
 
-    const tokenOpts = new Map();
-    tokenOpts.set("env", __ENV.ENVIRONMENT);
-    tokenOpts.set("ttl", 3600);
-    tokenOpts.set("scopes", "altinn:accessmanagement/authorizedparties.resourceowner");
-    const apClient = new AuthorizedPartiesClient(__ENV.BASE_URL, new EnterpriseTokenGenerator(tokenOpts));
-
     runErSyncTestcase(
-        "Add board member (MEDL)",
+        "Update contact information",
         prep,
         change,
         orgNr,
-        { "org is accessible in Register after MEDL added": (p) => p.partyType === "organization" },
         {
-            afterChange: () => {
-                group("Verify - new MEDL has access to org", () => {
-                    let verifiedParties = null;
-                    retry(
-                        () => {
-                            const parties = GetAuthorizedParties(apClient, "urn:altinn:person:identifier-no", NEW_MEDL.fnr, { includeAltinn2: false, includePartiesViaKeyRoles: true });
-                            if (!Array.isArray(parties)) return false;
-                            const hasAccess = parties.some((p) => p.organizationNumber === orgNr || p.orgNumber === orgNr);
-                            if (hasAccess) verifiedParties = parties;
-                            return hasAccess;
-                        },
-                        { retries: 15, intervalSeconds: 20, testscenario: "add-board-member - new MEDL access" },
-                    );
-                    check(verifiedParties, {
-                        [`new MEDL (${NEW_MEDL.fornavn} ${NEW_MEDL.slektsnavn}) has access to org`]: (p) => p !== null,
-                    });
-                });
-            },
+            "org telephoneNumber updated to 99887711": (p) => p.telephoneNumber === "99887711",
+            "org faxNumber updated to 99887712": (p) => p.faxNumber === "99887712",
+            "org emailAddress updated to contact-test-oppdatert@example.com": (p) => p.emailAddress === "contact-test-oppdatert@example.com",
+            "org internetAddress updated to http://oppdatert.example.com": (p) => p.internetAddress === "http://oppdatert.example.com",
         },
     );
 
@@ -165,13 +162,13 @@ function buildCleanupXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00410" mottaker="ALT" type="A" />
+            <head avsender="ER" dato="20260512" kjoerenr="00405" mottaker="ALT" type="A" />
             <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
+                <samendringer data="D" felttype="LEDE" endringstype="U" type="R">
+                    <rolleFoedselsnr>${LEDE.fnr}</rolleFoedselsnr>
+                </samendringer>
                 <samendringer data="D" felttype="DAGL" endringstype="U" type="R">
                     <rolleFoedselsnr>${DAGL.fnr}</rolleFoedselsnr>
-                </samendringer>
-                <samendringer data="D" felttype="MEDL" endringstype="U" type="R">
-                    <rolleFoedselsnr>${NEW_MEDL.fnr}</rolleFoedselsnr>
                 </samendringer>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
