+++ K6/api/tests/register/er-sync/testcase_8_add_frivillig_mva.js
@@ -5,29 +5,25 @@ import { RegisterApiClient } from "../../../../clients/authentication/index.js";
 import { SubmitErData } from "../../../building-blocks/register/index.js";
 
 /**
- * @file er-sync.js
- * @description Verifies that a change to NAVN (short name) in ER is correctly
+ * @file add-fmva.js
+ * @description Verifies that adding FMVA (Frivillig MVA-registrering) in ER is correctly
  * synced to Altinn Register.
  *
- * k6 run er-sync.js \
+ * k6 run add-fmva.js \
  *   -e ENVIRONMENT=at22 -e BASE_URL=https://platform.at22.altinn.cloud \
  *   -e SOAP_ER_USERNAME=<u> -e SOAP_ER_PASSWORD=<p> \
  *   -e REGISTER_SUBSCRIPTION_KEY=<key>
  *
- * @requires ENV.ENVIRONMENT                - Target environment (e.g. tt02, at22)
- * @requires ENV.BASE_URL                   - Base URL for the Register API
- * @requires ENV.SOAP_ER_USERNAME           - Username for the ER SOAP API
- * @requires ENV.SOAP_ER_PASSWORD           - Password for the ER SOAP API
- * @requires ENV.REGISTER_SUBSCRIPTION_KEY  - Subscription key for the Register API
+ * @see README.md
  */
 
 export const options = {
     scenarios: {
-        "change-org-short-name": { executor: "shared-iterations", exec: "nameShortChange", vus: 1, iterations: 1 },
+        "register-fmva": { executor: "shared-iterations", exec: "addFmva", vus: 1, iterations: 1 },
     },
 };
 
-const INNH = { fnr: "12864421537", fornavn: "Elise", slektsnavn: "Testperson" };
+const LEDE = { fnr: "02895823468", fornavn: "Anne", slektsnavn: "Testperson" };
 
 function buildPrepXml(orgNr) {
     return `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -38,26 +34,26 @@ function buildPrepXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00201" mottaker="ALT" type="A" />
-            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="ENK" hovedsakstype="N" undersakstype="NY" foersteOverfoering="J" datoFoedt="20210101" datoSistEndret="20260512">
+            <head avsender="ER" dato="20260512" kjoerenr="00223" mottaker="ALT" type="A" />
+            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="N" undersakstype="NY" foersteOverfoering="J" datoFoedt="20200101" datoSistEndret="20260512">
                 <infotype felttype="NAVN" endringstype="N">
-                    <navn1>NAVN CHANGE TEST ENK</navn1>
-                    <rednavn>NAVN CHANGE TEST ENK</rednavn>
+                    <navn1>FMVA ADD TEST AS</navn1>
+                    <rednavn>FMVA ADD TEST AS</rednavn>
                 </infotype>
                 <infotype felttype="FADR" endringstype="N">
-                    <postnr>1234</postnr>
+                    <postnr>0150</postnr>
                     <landkode>NO</landkode>
                     <kommunenr>0301</kommunenr>
-                    <adresse1>Testveien 1</adresse1>
+                    <adresse1>Testveien 10</adresse1>
                 </infotype>
-                <samendringer data="D" felttype="INNH" endringstype="N" type="R">
+                <samendringer data="D" felttype="LEDE" endringstype="N" type="R">
                     <rolleFratraadt>N</rolleFratraadt>
                     <rolleRekkefoelge>1</rolleRekkefoelge>
-                    <rolleFoedselsnr>${INNH.fnr}</rolleFoedselsnr>
-                    <fornavn>${INNH.fornavn}</fornavn>
-                    <slektsnavn>${INNH.slektsnavn}</slektsnavn>
-                    <postnr>1234</postnr>
-                    <adresse1>Testveien 29</adresse1>
+                    <rolleFoedselsnr>${LEDE.fnr}</rolleFoedselsnr>
+                    <fornavn>${LEDE.fornavn}</fornavn>
+                    <slektsnavn>${LEDE.slektsnavn}</slektsnavn>
+                    <postnr>0150</postnr>
+                    <adresse1>Testveien 11</adresse1>
                     <adresseLandkode>NO</adresseLandkode>
                     <personstatus>L</personstatus>
                 </samendringer>
@@ -73,7 +69,7 @@ export function setup() {
     return { orgNr: generateOrgNr() };
 }
 
-export function nameShortChange({ orgNr = generateOrgNr() } = {}) {
+export function addFmva({ orgNr = generateOrgNr() } = {}) {
     const prep = buildPrepXml(orgNr);
 
     const change = `<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ns="http://www.altinn.no/services/Register/ER/2013/06">
@@ -84,11 +80,10 @@ export function nameShortChange({ orgNr = generateOrgNr() } = {}) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00301" mottaker="ALT" type="A" />
-            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="ENK" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20210101" datoSistEndret="20260512">
-                <infotype felttype="NAVN" endringstype="N">
-                    <navn1>NAVN CHANGE TEST ENK OPPDATERT</navn1>
-                    <rednavn>NAVN CHANGE TEST ENK OPPDATERT</rednavn>
+            <head avsender="ER" dato="20260512" kjoerenr="00317" mottaker="ALT" type="A" />
+            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="NY" foersteOverfoering="N" datoFoedt="20260101" datoSistEndret="20260201">
+                <infotype felttype="FMVA" endringstype="N">
+                    <opplysning>OFVA</opplysning>
                 </infotype>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
@@ -98,11 +93,11 @@ export function nameShortChange({ orgNr = generateOrgNr() } = {}) {
 </soapenv:Envelope>`;
 
     runErSyncTestcase(
-        "Change organization short name",
+        "Register Frivillig MVA-registrering (FMVA)",
         prep,
         change,
         orgNr,
-        { "org name updated to NAVN CHANGE TEST ENK OPPDATERT": (p) => p.displayName === "NAVN CHANGE TEST ENK OPPDATERT" },
+        { "org is accessible in Register after FMVA added": (p) => p.partyType === "organization" },
     );
 
     group("Cleanup", () => {
@@ -123,10 +118,10 @@ function buildCleanupXml(orgNr) {
         <ns:systemPassword>${__ENV.SOAP_ER_PASSWORD}</ns:systemPassword>
             <ns:ERData><![CDATA[<?xml version="1.0" encoding="UTF-8"?>
         <batchAjourholdXML>
-            <head avsender="ER" dato="20260512" kjoerenr="00402" mottaker="ALT" type="A" />
-            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="ENK" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20210101" datoSistEndret="20260512">
-                <samendringer data="D" felttype="INNH" endringstype="U" type="R">
-                    <rolleFoedselsnr>${INNH.fnr}</rolleFoedselsnr>
+            <head avsender="ER" dato="20260512" kjoerenr="00403" mottaker="ALT" type="A" />
+            <enhet organisasjonsnummer="${orgNr}" organisasjonsform="AS" hovedsakstype="E" undersakstype="EN" foersteOverfoering="N" datoFoedt="20200101" datoSistEndret="20260512">
+                <samendringer data="D" felttype="LEDE" endringstype="U" type="R">
+                    <rolleFoedselsnr>${LEDE.fnr}</rolleFoedselsnr>
                 </samendringer>
             </enhet>
             <trai antallEnheter="1" avsender="ER" />
